Actually use doubles for armour damage reduction
This should help with rounding issues and make behaviour almost exactly the same as vanilla.

+++ src/main/java/net/modificationstation/stationloader/api/common/item/CustomArmourValue.java
@@ -3,5 +3,5 @@ package net.modificationstation.stationloader.api.common.item;
 import net.minecraft.entity.player.PlayerBase;
 
 public interface CustomArmourValue {
-    int modifyDamageDealt(PlayerBase playerBase, int armourSlot, int initialDamage, double currentAdjustedDamage);
+    double modifyDamageDealt(PlayerBase playerBase, int armourSlot, int initialDamage, double currentAdjustedDamage);
 }
